# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "papagaio-api",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/gettriggersconfig": {
            "get": {
                "security": [
                    {
                        "ApiKeyToken": []
                    }
                ],
                "description": "Get trigger timers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Return time triggers",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/dto.ConfigTriggersDto"
                        }
                    }
                }
            }
        },
...
        "/restarttriggers": {
            "post": {
                "security": [
                    {
                        "ApiKeyToken": []
                    }
                ],
                "description": "Restart timers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "restart triggers",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "?restartAll",
                        "name": "restartAll",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "?restartorganizationsynktrigger",
                        "name": "restartorganizationsynktrigger",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "?restartRunsFailedDiscoveryTrigger",
                        "name": "restartRunsFailedDiscoveryTrigger",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "?restartUsersSynkTrigger",
                        "name": "restartUsersSynkTrigger",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/savetriggersconfig": {
            "post": {
                "security": [
                    {
                        "ApiKeyToken": []
                    }
                ],
                "description": "Save trigger timers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "Save time triggers",
                "parameters": [
                    {
                        "description": "Config triggers",
                        "name": "configTriggersDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ConfigTriggersDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        },
        "/triggersstatus": {
            "get": {
                "security": [
                    {
                        "ApiKeyToken": []
                    }
                ],
                "description": "Get triggers status",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Triggers"
                ],
                "summary": "get triggers status",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.ConfigTriggersDto": {
            "type": "object",
            "properties": {
                "organizationsTriggerTime": {
                    "type": "integer"
                },
                "runFailedTriggerTime": {
                    "type": "integer"
                },
                "usersTriggerTime": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}